Reorganize client modules
