+++ examples/server.rs
@@ -22,7 +22,7 @@ use futures::sync::mpsc;
 use hyper::header;
 use hyper::server::{Http, Request, Response, Service};
 use hyper_tls::HttpsConnector;
-use petronel::{ClientBuilder, Petronel, Subscriber, Subscription, Token};
+use petronel::{Client, ClientBuilder, Subscriber, Subscription, Token};
 use petronel::error::*;
 use petronel::model::Message;
 use regex::Regex;
@@ -53,24 +53,25 @@ quick_main!(|| -> Result<()> {
     let listener = tokio_core::net::TcpListener::bind(&bind_address, &handle)
         .chain_err(|| "failed to bind TCP listener")?;
 
-    let client = hyper::Client::configure()
+    let hyper_client = hyper::Client::configure()
         .connector(HttpsConnector::new(4, &handle).chain_err(|| "HTTPS error")?)
         .build(&handle);
 
-    let (petronel, petronel_worker) = ClientBuilder::from_hyper_client(&client, &token)
-        .with_history_size(10)
-        .with_subscriber::<Sender>()
-        .map_message(|msg| match msg {
-            Message::Heartbeat => "\n".into(),
-            other => {
-                let mut bytes = serde_json::to_vec(&other).unwrap();
-                bytes.push(b'\n');
-                bytes.into()
-            }
-        })
-        .build();
+    let (petronel_client, petronel_worker) =
+        ClientBuilder::from_hyper_client(&hyper_client, &token)
+            .with_history_size(10)
+            .with_subscriber::<Sender>()
+            .map_message(|msg| match msg {
+                Message::Heartbeat => "\n".into(),
+                other => {
+                    let mut bytes = serde_json::to_vec(&other).unwrap();
+                    bytes.push(b'\n');
+                    bytes.into()
+                }
+            })
+            .build();
 
-    let petronel_server = PetronelServer(petronel.clone());
+    let petronel_server = PetronelServer(petronel_client.clone());
 
     println!("Listening on {}", bind_address);
 
@@ -86,7 +87,7 @@ quick_main!(|| -> Result<()> {
     // Send heartbeat every 30 seconds
     let heartbeat = Interval::new(Duration::new(30, 0), &core.handle())
         .chain_err(|| "failed to create Interval")?
-        .for_each(move |_| Ok(petronel.heartbeat()))
+        .for_each(move |_| Ok(petronel_client.heartbeat()))
         .then(|r| r.chain_err(|| "heartbeat failed"));
 
     core.run(server.join3(petronel_worker, heartbeat))
@@ -108,7 +109,7 @@ impl Subscriber for Sender {
     }
 }
 
-struct PetronelServer(Petronel<Sender>);
+struct PetronelServer(Client<Sender>);
 
 impl Clone for PetronelServer {
     fn clone(&self) -> Self {
