+++ src/client/client.rs
@@ -1,15 +1,13 @@
+use super::{Event, RaidBossEntry, Subscription};
 use broadcast::{Broadcast, Subscriber};
 use circular_buffer::CircularBuffer;
 use error::*;
 use futures::{Async, Future, Poll, Stream};
 use futures::stream::{Map, OrElse, Select};
 use futures::unsync::{mpsc, oneshot};
-use hyper::Client;
-use hyper::client::Connect;
 use id_pool::{Id as SubId, IdPool};
-use image_hash::{self, BossImageHash, HyperImageHasher, ImageHash, ImageHashReceiver,
-                 ImageHashSender, ImageHasher};
-use model::{BossLevel, BossName, DateTime, Message, RaidBoss, RaidTweet};
+use image_hash::{BossImageHash, ImageHash, ImageHashReceiver, ImageHashSender, ImageHasher};
+use model::{BossLevel, BossName, Message, RaidBoss, RaidTweet};
 use raid::RaidInfo;
 use std::collections::{HashMap, HashSet};
 use std::collections::hash_map::Entry;
@@ -17,43 +15,6 @@ use std::iter::FromIterator;
 use std::sync::Arc;
 
 const DEFAULT_BOSS_LEVEL: BossLevel = 0;
-
-pub(crate) struct RaidBossEntry<Sub> {
-    boss: RaidBoss,
-    last_seen: DateTime,
-    image_hash: Option<ImageHash>,
-    recent_tweets: CircularBuffer<Arc<RaidTweet>>,
-    broadcast: Broadcast<SubId, Sub>,
-}
-
-#[derive(Debug)]
-pub(crate) enum Event<Sub> {
-    NewRaidInfo(RaidInfo),
-    NewImageHash {
-        boss_name: BossName,
-        image_hash: ImageHash,
-    },
-
-    Follow { id: SubId, boss_name: BossName },
-    Unfollow { id: SubId, boss_name: BossName },
-
-    Subscribe {
-        subscriber: Sub,
-        petronel: Petronel<Sub>,
-        sender: oneshot::Sender<Subscription<Sub>>,
-    },
-    Unsubscribe(SubId),
-
-    GetBosses(oneshot::Sender<Vec<RaidBoss>>),
-    GetRecentTweets {
-        boss_name: BossName,
-        sender: oneshot::Sender<Vec<Arc<RaidTweet>>>,
-    },
-
-    Heartbeat,
-
-    ReadError,
-}
 
 pub struct AsyncResult<T>(oneshot::Receiver<T>);
 impl<T> Future for AsyncResult<T> {
@@ -66,67 +27,16 @@ impl<T> Future for AsyncResult<T> {
 }
 
 #[derive(Debug)]
-pub struct Petronel<Sub>(pub(crate) mpsc::UnboundedSender<Event<Sub>>);
+pub struct Client<Sub>(pub(crate) mpsc::UnboundedSender<Event<Sub>>);
 
-impl<Sub> Clone for Petronel<Sub> {
+impl<Sub> Clone for Client<Sub> {
     fn clone(&self) -> Self {
-        Petronel(self.0.clone())
-    }
-}
-
-// TODO: Figure out if there is a way to do this without owning `Petronel`
-#[must_use = "Subscriptions are cancelled when they go out of scope"]
-#[derive(Debug)]
-pub struct Subscription<Sub> {
-    id: SubId,
-    following: HashSet<BossName>,
-    petronel: Petronel<Sub>,
-}
-
-impl<Sub> Subscription<Sub> {
-    pub fn follow<B>(&mut self, boss_name: B)
-    where
-        B: Into<BossName>,
-    {
-        let name = boss_name.into();
-        self.following.insert(name.clone());
-        self.petronel.follow(self.id.clone(), name);
-    }
-
-    pub fn unfollow<B>(&mut self, boss_name: B)
-    where
-        B: Into<BossName>,
-    {
-        let name = boss_name.into();
-        self.following.remove(&name);
-        self.petronel.unfollow(self.id.clone(), name);
-    }
-
-    #[inline]
-    pub fn unsubscribe(self) {
-        self.non_consuming_unsubscribe()
-    }
-
-    // This is needed for the Drop implementation
-    fn non_consuming_unsubscribe(&self) {
-        self.petronel.unsubscribe(self.id.clone())
-    }
-}
-
-impl<Sub> Drop for Subscription<Sub> {
-    fn drop(&mut self) {
-        let mut following = ::std::mem::replace(&mut self.following, HashSet::with_capacity(0));
-
-        for boss_name in following.drain() {
-            self.unfollow(boss_name);
-        }
-
-        self.non_consuming_unsubscribe();
+        Client(self.0.clone())
     }
 }
 
 
-impl<Sub> Petronel<Sub> {
+impl<Sub> Client<Sub> {
     fn send(&self, event: Event<Sub>) {
         let _ = mpsc::UnboundedSender::send(&self.0, event);
     }
@@ -145,20 +55,20 @@ impl<Sub> Petronel<Sub> {
             Event::Subscribe {
                 subscriber,
                 sender,
-                petronel: self.clone(),
+                client: self.clone(),
             }
         })
     }
 
-    fn unsubscribe(&self, id: SubId) {
+    pub(crate) fn unsubscribe(&self, id: SubId) {
         self.send(Event::Unsubscribe(id));
     }
 
-    fn follow(&self, id: SubId, boss_name: BossName) {
+    pub(crate) fn follow(&self, id: SubId, boss_name: BossName) {
         self.send(Event::Follow { id, boss_name });
     }
 
-    fn unfollow(&self, id: SubId, boss_name: BossName) {
+    pub(crate) fn unfollow(&self, id: SubId, boss_name: BossName) {
         self.send(Event::Unfollow { id, boss_name });
     }
 
@@ -184,7 +94,7 @@ impl<Sub> Petronel<Sub> {
 }
 
 #[must_use = "futures do nothing unless polled"]
-pub struct PetronelFuture<H, S, Sub, F>
+pub struct ClientWorker<H, S, Sub, F>
 where
     H: ImageHasher,
 {
@@ -208,79 +118,26 @@ where
     pub(crate) map_message: F,
 }
 
-impl<Sub> Petronel<Sub> {
-    fn events_read_error(_: ()) -> Result<Event<Sub>> {
-        Ok(Event::ReadError)
-    }
-
-    fn boss_image_hash_to_event(msg: BossImageHash) -> Event<Sub> {
-        Event::NewImageHash {
-            boss_name: msg.boss_name,
-            image_hash: msg.image_hash,
-        }
-    }
-
-    // TODO: Builder
-    pub fn from_stream<'a, C, S, F>(
-        stream: S,
-        tweet_history_size: usize,
-        hyper: &'a Client<C>,
-        map_message: F,
-    ) -> (Self, PetronelFuture<HyperImageHasher<'a, C>, S, Sub, F>)
-    where
-        C: Connect,
-        S: Stream<Item = RaidInfo, Error = Error>,
-        Sub: Subscriber,
-        F: Fn(Message) -> Sub::Item,
-    {
-        let (tx, rx) = mpsc::unbounded();
-
-        let stream_events = stream.map(Event::NewRaidInfo as fn(RaidInfo) -> Event<Sub>);
-        let rx = rx.or_else(Self::events_read_error as fn(()) -> Result<Event<Sub>>);
-
-        // TODO: Configurable
-        let hasher = HyperImageHasher(hyper);
-        let (hash_requester, hash_receiver) = image_hash::channel(hasher, 10);
-        let hash_events = hash_receiver.map(
-            Self::boss_image_hash_to_event as
-                fn(BossImageHash) -> Event<Sub>,
-        );
-
-        let future = PetronelFuture {
-            hash_requester,
-            id_pool: IdPool::new(),
-            events: stream_events.select(rx.select(hash_events)),
-            bosses: HashMap::new(),
-            tweet_history_size,
-            requested_bosses: HashMap::new(),
-            subscribers: Broadcast::new(),
-            map_message,
-        };
-
-        (Petronel(tx), future)
-    }
-}
-
-impl<H, S, Sub, F> PetronelFuture<H, S, Sub, F>
+impl<H, S, Sub, F> ClientWorker<H, S, Sub, F>
 where
     H: ImageHasher,
     Sub: Subscriber + Clone,
     F: Fn(Message) -> Sub::Item,
 {
     fn handle_event(&mut self, event: Event<Sub>) {
-        use self::Event::*;
+        use super::Event::*;
 
         match event {
             Subscribe {
                 subscriber,
                 sender,
-                petronel,
+                client,
             } => {
                 let id = self.subscribe(subscriber);
                 let _ = sender.send(Subscription {
                     id,
                     following: HashSet::new(),
-                    petronel,
+                    client,
                 });
             }
             Unsubscribe(id) => {
@@ -474,7 +331,7 @@ where
     }
 }
 
-impl<H, S, Sub, F> Future for PetronelFuture<H, S, Sub, F>
+impl<H, S, Sub, F> Future for ClientWorker<H, S, Sub, F>
 where
     H: ImageHasher,
     S: Stream<Item = RaidInfo, Error = Error>,
